Add a new m_loadedBuffer string member variable that holds the string buffer that's saved in SaveBufferToPersistentStorage() and make it so that it's loaded into m_loadedBuffer in LoadBufferFromPersistentStorage()
Fix SaveBufferToPersistentStorage() to have it convert the AZStd::string
object that's passed in into a character array / C string. It was found
that when dynamically converting the AZStd::string a length of >=23
characters (>=24 characters when including the null terminating
character) caused the first 8 characters / bytes in the saved text file
to be bogus, which also had issues when loading.

Signed-off-by: Eric Kuzmenko <eric.n.kuzmenko@gmail.com>